split server.go
