+++ services/user/internal/server/server.go
@@ -1,32 +1,13 @@
 package server
 
 import (
-	"bytes"
-	"context"
-	"crypto/rand"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/base64"
-	"errors"
-	"fmt"
-	"net/url"
-	"os"
-	"reflect"
-	"slices"
-	"strings"
 	"time"
 
-	"github.com/RAF-SI-2025/Banka-3-Backend/services/user/internal/model"
 	"github.com/RAF-SI-2025/Banka-3-Backend/services/user/internal/repo"
-	"github.com/golang-jwt/jwt/v5"
 	"github.com/redis/go-redis/v9"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/credentials/insecure"
-	"google.golang.org/grpc/status"
 	"gorm.io/gorm"
 
-	"github.com/RAF-SI-2025/Banka-3-Backend/pkg/logger"
 	notificationpb "github.com/RAF-SI-2025/Banka-3-Backend/pkg/proto/notification"
 	userpb "github.com/RAF-SI-2025/Banka-3-Backend/pkg/proto/user"
 )
@@ -70,977 +51,4 @@ func NewServer(accessJwtSecret string, refreshJwtSecret string, conn *Connection
 			Gorm:     conn.Gorm,
 		},
 	}
-}
-
-func generateSalt() ([]byte, error) {
-	salt := make([]byte, 16)
-	_, err := rand.Read(salt)
-	if err != nil {
-		return nil, err
-	}
-	return salt, nil
-}
-
-func HashPassword(password string, salt []byte) []byte {
-	hashed := sha256.New()
-	hashed.Write(salt)
-	hashed.Write([]byte(password))
-	return hashed.Sum(nil)
-}
-
-func (s *Server) GetEmployeeByEmail(_ context.Context, req *userpb.GetUserByEmailRequest) (*userpb.GetEmployeeResponse, error) {
-	resp, err := s.repo.GetEmployeeByAttribute("email", req.Email)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, status.Error(codes.NotFound, "employee not found")
-		}
-		return nil, status.Error(codes.Internal, "failed to get employee")
-	}
-	return resp.ToProtobuf(), nil
-}
-
-func (s *Server) GetEmployeeById(_ context.Context, req *userpb.GetUserByIdRequest) (*userpb.GetEmployeeResponse, error) {
-	resp, err := s.repo.GetEmployeeByAttribute("id", req.Id)
-
-	if err != nil {
-		return nil, err
-	}
-	return resp.ToProtobuf(), nil
-}
-
-func (s *Server) GetClientByEmail(_ context.Context, req *userpb.GetUserByEmailRequest) (*userpb.GetClientResponse, error) {
-	resp, err := s.repo.GetClientByAttribute("email", req.Email)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, status.Error(codes.NotFound, "employee not found")
-		}
-		return nil, status.Error(codes.Internal, "failed to get employee")
-	}
-	return resp.ToProtobuf(), nil
-}
-
-func (s *Server) GetClientById(_ context.Context, req *userpb.GetUserByIdRequest) (*userpb.GetClientResponse, error) {
-	resp, err := s.repo.GetClientByAttribute("id", req.Id)
-	if err != nil {
-		return nil, err
-	}
-	return resp.ToProtobuf(), nil
-}
-
-func (s *Server) DeleteEmployee(_ context.Context, req *userpb.DeleteEmployeeRequest) (*userpb.DeleteEmployeeResponse, error) {
-	err := s.repo.DeleteEmployee(model.Employee{
-		Id: uint64(req.Id),
-	})
-
-	if err != nil {
-		if errors.Is(err, repo.ErrEmployeeNotFound) {
-			return nil, status.Error(codes.NotFound, "employee not found")
-		}
-		return nil, err
-	}
-	return &userpb.DeleteEmployeeResponse{Success: true}, nil
-}
-
-func (s *Server) GetEmployees(ctx context.Context, req *userpb.GetEmployeesRequest) (*userpb.GetEmployeesResponse, error) {
-	map_func := func(emp model.Employee) *userpb.GetEmployeesResponse_Employee {
-		return &userpb.GetEmployeesResponse_Employee{
-			Id:           int64(emp.Id),
-			FirstName:    emp.First_name,
-			LastName:     emp.Last_name,
-			Email:        emp.Email,
-			Position:     emp.Position,
-			PhoneNumber:  emp.Phone_number,
-			Active:       emp.Active,
-			Limit:        emp.Limit,
-			UsedLimit:    emp.Used_limit,
-			NeedApproval: emp.Need_approval,
-		}
-	}
-	restrictions := repo.UserRestrictions{"first_name": req.FirstName, "last_name": req.LastName, "email": req.Email, "position": req.Position}
-
-	employees, err := s.repo.GetAllEmployees(restrictions)
-	if err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "error retrieving employees", "err", err)
-		return nil, status.Error(codes.Internal, "Failed to retrieve employees")
-	}
-
-	var employee_responses []*userpb.GetEmployeesResponse_Employee
-	for _, emp := range employees {
-		employee_responses = append(employee_responses, map_func(emp))
-	}
-
-	return &userpb.GetEmployeesResponse{Employees: employee_responses}, nil
-}
-
-func (s *Server) UpdateEmployee(ctx context.Context, req *userpb.UpdateEmployeeRequest) (*userpb.GetEmployeeResponse, error) {
-	existing, existingErr := s.repo.GetEmployeeByAttribute("id", req.Id)
-
-	// Spec p.38: an admin may not edit another admin. Self-edits are allowed.
-	if existingErr == nil && existing != nil {
-		existingPerms := permissionSet(existing.Permissions)
-		if _, targetIsAdmin := existingPerms["admin"]; targetIsAdmin &&
-			!strings.EqualFold(strings.TrimSpace(req.CallerEmail), existing.Email) {
-			return nil, status.Error(codes.PermissionDenied, "admins cannot edit other admins")
-		}
-	}
-
-	if !req.Active {
-		if existingErr == nil && existing != nil {
-			for _, p := range existing.Permissions {
-				if p.Name == "admin" {
-					return nil, status.Error(codes.PermissionDenied, "cannot deactivate an admin")
-				}
-			}
-		}
-	}
-
-	// Spec p.38: every admin is also a supervisor. Granting admin implicitly
-	// grants supervisor; revoking admin leaves supervisor untouched.
-	if req.Permissions != nil {
-		req.Permissions = EnsureAdminImpliesSupervisor(req.Permissions)
-	}
-
-	// Only admins may grant or revoke the `agent` / `supervisor` permissions.
-	if req.Permissions != nil && existingErr == nil && existing != nil {
-		oldSet := permissionSet(existing.Permissions)
-		newSet := NamesToSet(req.Permissions)
-		if TogglesTradingRole(oldSet, newSet) {
-			if !callerIsAdmin(ctx, s, req.CallerEmail) {
-				return nil, status.Error(codes.PermissionDenied, "only admins may grant or revoke agent/supervisor permissions")
-			}
-		}
-	}
-
-	var permissions []model.Permission
-	for _, perm := range req.Permissions {
-		// yes these are invalid. i don't care
-		permissions = append(permissions, model.Permission{Id: 0, Name: perm})
-	}
-
-	emp := model.Employee{
-		Last_name:    req.LastName,
-		Gender:       req.Gender,
-		Phone_number: req.PhoneNumber,
-		Address:      req.Address,
-		Position:     req.Position,
-		Department:   req.Department,
-		Active:       req.Active,
-		Id:           uint64(req.Id),
-		Updated_at:   time.Now(),
-		Permissions:  permissions,
-	}
-
-	updated, err := s.repo.UpdateEmployee(emp)
-	if err != nil {
-		if errors.Is(err, repo.ErrEmployeeNotFound) {
-			return nil, status.Error(codes.NotFound, "Employee not found")
-		}
-		if errors.Is(err, repo.ErrUnknownPermission) {
-			return nil, status.Error(codes.NotFound, "Uknown permissions")
-		}
-		return nil, status.Error(codes.Internal, "Messed something up in UpdateEmployee_ in repo")
-	}
-
-	// Sync session: deactivation deletes session, otherwise update permissions
-	if !req.Active {
-		_ = s.DeleteSession(ctx, updated.Email)
-	} else {
-		permNames := make([]string, len(updated.Permissions))
-		for i, p := range updated.Permissions {
-			permNames[i] = p.Name
-		}
-		_ = s.UpdateSessionPermissions(ctx, updated.Email, "employee", permNames)
-	}
-
-	return updated.ToProtobuf(), nil
-
-}
-
-// UpdateEmployeeTradingLimit sets an agent's daily trading limit and/or used_limit.
-// Only admins and supervisors may call this. The caller is identified by CallerEmail
-// (forwarded from the gateway).
-func (s *Server) UpdateEmployeeTradingLimit(ctx context.Context, req *userpb.UpdateEmployeeTradingLimitRequest) (*userpb.GetEmployeeResponse, error) {
-	if req.Id <= 0 {
-		return nil, status.Error(codes.InvalidArgument, "id must be greater than zero")
-	}
-	if req.Limit == nil && req.UsedLimit == nil {
-		return nil, status.Error(codes.InvalidArgument, "limit or used_limit must be provided")
-	}
-	if req.Limit != nil && *req.Limit < 0 {
-		return nil, status.Error(codes.InvalidArgument, "limit must be non-negative")
-	}
-	if req.UsedLimit != nil && *req.UsedLimit < 0 {
-		return nil, status.Error(codes.InvalidArgument, "used_limit must be non-negative")
-	}
-
-	if !callerCanManageLimits(ctx, s, req.CallerEmail) {
-		return nil, status.Error(codes.PermissionDenied, "only admins and supervisors may change an agent's limit")
-	}
-
-	target, err := s.repo.GetEmployeeByAttribute("id", req.Id)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, status.Error(codes.NotFound, "employee not found")
-		}
-		return nil, status.Error(codes.Internal, "failed to load employee")
-	}
-
-	updates := map[string]any{"updated_at": time.Now()}
-	if req.Limit != nil {
-		updates["limit"] = *req.Limit
-	}
-	if req.UsedLimit != nil {
-		updates["used_limit"] = *req.UsedLimit
-	}
-
-	if err := s.db_gorm.Model(&model.Employee{}).Where("id = ?", target.Id).Updates(updates).Error; err != nil {
-		return nil, status.Error(codes.Internal, "failed to update trading limit")
-	}
-
-	reloaded, err := s.repo.GetEmployeeByAttribute("id", req.Id)
-	if err != nil {
-		return nil, status.Error(codes.Internal, "failed to reload employee")
-	}
-	return reloaded.ToProtobuf(), nil
-}
-
-// UpdateEmployeeNeedApproval flips the need_approval flag on an employee. Admins
-// and supervisors only — caller is identified by CallerEmail.
-func (s *Server) UpdateEmployeeNeedApproval(ctx context.Context, req *userpb.UpdateEmployeeNeedApprovalRequest) (*userpb.GetEmployeeResponse, error) {
-	if req.Id <= 0 {
-		return nil, status.Error(codes.InvalidArgument, "id must be greater than zero")
-	}
-	if !callerCanManageLimits(ctx, s, req.CallerEmail) {
-		return nil, status.Error(codes.PermissionDenied, "only admins and supervisors may change need_approval")
-	}
-
-	target, err := s.repo.GetEmployeeByAttribute("id", req.Id)
-	if err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, status.Error(codes.NotFound, "employee not found")
-		}
-		return nil, status.Error(codes.Internal, "failed to load employee")
-	}
-
-	if err := s.db_gorm.Model(&model.Employee{}).Where("id = ?", target.Id).Updates(map[string]any{
-		"need_approval": req.NeedApproval,
-		"updated_at":    time.Now(),
-	}).Error; err != nil {
-		return nil, status.Error(codes.Internal, "failed to update need_approval")
-	}
-
-	reloaded, err := s.repo.GetEmployeeByAttribute("id", req.Id)
-	if err != nil {
-		return nil, status.Error(codes.Internal, "failed to reload employee")
-	}
-	return reloaded.ToProtobuf(), nil
-}
-
-// GetActuaries returns employees that hold the `agent` permission, with the same
-// filter set as GetEmployees. Spec page 39 — supervisor portal.
-func (s *Server) GetActuaries(ctx context.Context, req *userpb.GetEmployeesRequest) (*userpb.GetEmployeesResponse, error) {
-	restrictions := repo.UserRestrictions{
-		"first_name": req.FirstName,
-		"last_name":  req.LastName,
-		"email":      req.Email,
-		"position":   req.Position,
-	}
-
-	employees, err := s.repo.GetAllEmployees(restrictions)
-	if err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "error retrieving actuaries", "err", err)
-		return nil, status.Error(codes.Internal, "Failed to retrieve actuaries")
-	}
-
-	out := make([]*userpb.GetEmployeesResponse_Employee, 0, len(employees))
-	for _, emp := range employees {
-		isAgent := false
-		for _, p := range emp.Permissions {
-			if p.Name == "agent" {
-				isAgent = true
-				break
-			}
-		}
-		if !isAgent {
-			continue
-		}
-		out = append(out, &userpb.GetEmployeesResponse_Employee{
-			Id:           int64(emp.Id),
-			FirstName:    emp.First_name,
-			LastName:     emp.Last_name,
-			Email:        emp.Email,
-			Position:     emp.Position,
-			PhoneNumber:  emp.Phone_number,
-			Active:       emp.Active,
-			Limit:        emp.Limit,
-			UsedLimit:    emp.Used_limit,
-			NeedApproval: emp.Need_approval,
-		})
-	}
-	return &userpb.GetEmployeesResponse{Employees: out}, nil
-}
-
-// permissionSet converts a list of Permission rows to a string set for easy membership tests.
-func permissionSet(perms []model.Permission) map[string]struct{} {
-	out := make(map[string]struct{}, len(perms))
-	for _, p := range perms {
-		out[p.Name] = struct{}{}
-	}
-	return out
-}
-
-func NamesToSet(names []string) map[string]struct{} {
-	out := make(map[string]struct{}, len(names))
-	for _, n := range names {
-		out[n] = struct{}{}
-	}
-	return out
-}
-
-// EnsureAdminImpliesSupervisor returns perms with "supervisor" appended when
-// "admin" is present but "supervisor" is not (spec p.38: admin is-a supervisor).
-// Idempotent: calling twice yields the same result.
-func EnsureAdminImpliesSupervisor(perms []string) []string {
-	set := NamesToSet(perms)
-	if _, hasAdmin := set["admin"]; !hasAdmin {
-		return perms
-	}
-	if _, hasSup := set["supervisor"]; hasSup {
-		return perms
-	}
-	return append(perms, "supervisor")
-}
-
-// TogglesTradingRole reports whether the `agent` or `supervisor` membership differs
-// between the old and new permission sets.
-func TogglesTradingRole(oldSet, newSet map[string]struct{}) bool {
-	for _, perm := range []string{"agent", "supervisor"} {
-		_, inOld := oldSet[perm]
-		_, inNew := newSet[perm]
-		if inOld != inNew {
-			return true
-		}
-	}
-	return false
-}
-
-func callerIsAdmin(_ context.Context, s *Server, callerEmail string) bool {
-	if strings.TrimSpace(callerEmail) == "" {
-		return false
-	}
-	caller, err := s.repo.GetEmployeeByAttribute("email", callerEmail)
-	if err != nil || caller == nil {
-		return false
-	}
-	for _, p := range caller.Permissions {
-		if p.Name == "admin" {
-			return true
-		}
-	}
-	return false
-}
-
-func callerCanManageLimits(_ context.Context, s *Server, callerEmail string) bool {
-	if strings.TrimSpace(callerEmail) == "" {
-		return false
-	}
-	caller, err := s.repo.GetEmployeeByAttribute("email", callerEmail)
-	if err != nil || caller == nil {
-		return false
-	}
-	for _, p := range caller.Permissions {
-		if p.Name == "admin" || p.Name == "supervisor" {
-			return true
-		}
-	}
-	return false
-}
-
-func (s *Server) GetClients(ctx context.Context, req *userpb.GetClientsRequest) (*userpb.GetClientsResponse, error) {
-
-	clients, err := s.repo.GetAllClients(repo.UserRestrictions{"first_name": strings.TrimSpace(req.FirstName), "last_name": strings.TrimSpace(req.LastName), "email": strings.TrimSpace(req.Email)})
-
-	if err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "error retrieving clients", "err", err)
-		return nil, status.Error(codes.Internal, "Failed to retrieve clients")
-	}
-
-	var clientResponses []*userpb.Client
-	for _, client := range clients {
-		clientResponses = append(clientResponses, client.ToProtobuff())
-	}
-
-	return &userpb.GetClientsResponse{Clients: clientResponses}, nil
-}
-
-func (s *Server) UpdateClient(ctx context.Context, req *userpb.UpdateClientRequest) (*userpb.UpdateClientResponse, error) {
-	if req.Id <= 0 {
-		return nil, status.Error(codes.InvalidArgument, "id must be greater than zero")
-	}
-	if strings.TrimSpace(req.Gender) != "" && req.Gender != "M" && req.Gender != "F" {
-		return nil, status.Error(codes.InvalidArgument, "Gender must be one of M or F")
-	}
-	client := model.Client{
-		Id:           uint64(req.Id),
-		First_name:   req.FirstName,
-		Last_name:    req.LastName,
-		Gender:       req.Gender,
-		Email:        req.Email,
-		Phone_number: req.PhoneNumber,
-		Address:      req.Address,
-	}
-
-	// I hope any potential reader of this has as much fun reading it as I had Implementing it.
-	ref := reflect.ValueOf(&client).Elem()
-	for i := 0; i < ref.NumField(); i++ {
-		field := ref.Field(i)
-		if field.Type() == reflect.TypeFor[string]() {
-			if !field.CanSet() {
-				logger.FromContext(ctx).ErrorContext(ctx, "cannot set the value of struct field")
-				// This need not be an error, but it will also probably
-				// never happen
-				return nil, status.Error(codes.Internal, "client update failed")
-			}
-			field.SetString(strings.TrimSpace(field.String()))
-		}
-
-	}
-
-	if req.DateOfBirth != 0 {
-		client.Date_of_birth = time.Unix(req.DateOfBirth, 0)
-	}
-
-	_, err := s.repo.UpdateClient(client)
-	if err != nil {
-		switch {
-		case errors.Is(err, repo.ErrClientNotFound):
-			return nil, status.Error(codes.NotFound, "client not found")
-		case errors.Is(err, repo.ErrClientEmailExists):
-			return nil, status.Error(codes.AlreadyExists, "client with that email already exists")
-		case errors.Is(err, repo.ErrClientNoFieldsToUpdate):
-			return nil, status.Error(codes.InvalidArgument, "no fields to update")
-		default:
-			return nil, status.Error(codes.Internal, "client update failed")
-		}
-	}
-
-	return &userpb.UpdateClientResponse{Valid: true, Response: "Client updated"}, nil
-}
-
-func (s *Server) GenerateRefreshToken(email string) (string, error) {
-	now := time.Now()
-	claims := jwt.RegisteredClaims{
-		Subject:   email,
-		ExpiresAt: jwt.NewNumericDate(now.Add(24 * time.Hour * 7)),
-		IssuedAt:  jwt.NewNumericDate(now),
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	return token.SignedString([]byte(s.refreshJwtSecret))
-}
-
-func (s *Server) GenerateAccessToken(email string) (string, error) {
-	now := time.Now()
-	claims := jwt.RegisteredClaims{
-		Subject:   email,
-		ExpiresAt: jwt.NewNumericDate(now.Add(15 * time.Minute)),
-		IssuedAt:  jwt.NewNumericDate(now),
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	return token.SignedString([]byte(s.accessJwtSecret))
-}
-
-func validateJWTToken(tokenString, secret string) (*userpb.ValidateTokenResponse, error) {
-	claims := &jwt.RegisteredClaims{}
-	token, err := jwt.ParseWithClaims(tokenString, claims, func(t *jwt.Token) (any, error) {
-		return []byte(secret), nil
-	})
-
-	if err != nil || !token.Valid {
-		return nil, status.Error(codes.Unauthenticated, "invalid token")
-	}
-
-	sub, err := claims.GetSubject()
-	if err != nil {
-		return nil, err
-	}
-	exp, err := claims.GetExpirationTime()
-	if err != nil {
-		return nil, err
-	}
-	iat, err := claims.GetIssuedAt()
-	if err != nil {
-		return nil, err
-	}
-
-	return &userpb.ValidateTokenResponse{
-		Sub: sub,
-		Exp: exp.Unix(),
-		Iat: iat.Unix(),
-	}, nil
-}
-
-func (s *Server) ValidateRefreshToken(_ context.Context, req *userpb.ValidateTokenRequest) (*userpb.ValidateTokenResponse, error) {
-	return validateJWTToken(req.Token, s.refreshJwtSecret)
-}
-
-func (s *Server) ValidateAccessToken(ctx context.Context, req *userpb.ValidateTokenRequest) (*userpb.ValidateTokenResponse, error) {
-	resp, err := validateJWTToken(req.Token, s.accessJwtSecret)
-	if err != nil {
-		return nil, err
-	}
-
-	session, err := s.GetSession(ctx, resp.Sub)
-	if err != nil {
-		return nil, status.Error(codes.Unavailable, "session store unavailable")
-	}
-	if session == nil {
-		return nil, status.Error(codes.Unauthenticated, "no active session")
-	}
-	if !session.Active {
-		return nil, status.Error(codes.Unauthenticated, "account deactivated")
-	}
-
-	return resp, nil
-}
-
-func (s *Server) GetUserPermissions(ctx context.Context, req *userpb.GetUserPermissionsRequest) (*userpb.GetUserPermissionsResponse, error) {
-	session, err := s.GetSession(ctx, req.Email)
-	if err != nil {
-		return nil, status.Error(codes.Unavailable, "session store unavailable")
-	}
-	if session == nil {
-		return nil, status.Error(codes.Unauthenticated, "no active session")
-	}
-
-	return &userpb.GetUserPermissionsResponse{
-		Role:        session.Role,
-		Permissions: session.Permissions,
-	}, nil
-}
-
-func (s *Server) Refresh(ctx context.Context, req *userpb.RefreshRequest) (*userpb.RefreshResponse, error) {
-	token, err := validateJWTToken(req.RefreshToken, s.refreshJwtSecret)
-	if err != nil {
-		return nil, err
-	}
-	email := token.Sub
-
-	newSignedToken, err := s.GenerateRefreshToken(email)
-	if err != nil {
-		return nil, fmt.Errorf("generating refresh token: %w", err)
-	}
-
-	role, permissions, active := s.getRoleAndPermissions(email)
-
-	if !active {
-		return nil, status.Error(codes.Unauthenticated, "account deactivated")
-	}
-
-	newAccessToken, err := s.GenerateAccessToken(email)
-	if err != nil {
-		return nil, fmt.Errorf("generating access token: %w", err)
-	}
-
-	newParsed, _, err := jwt.NewParser().ParseUnverified(newSignedToken, &jwt.RegisteredClaims{})
-	if err != nil {
-		return nil, fmt.Errorf("parsing new token: %w", err)
-	}
-	newExpiry, err := newParsed.Claims.GetExpirationTime()
-	if err != nil {
-		return nil, fmt.Errorf("getting expiry: %w", err)
-	}
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return nil, fmt.Errorf("starting transaction: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	err = s.repo.RotateRefreshToken(tx, email, HashValue(req.RefreshToken), HashValue(newSignedToken), newExpiry.Time)
-	if err != nil {
-		return nil, status.Error(codes.Unauthenticated, "wrong token")
-	}
-
-	if err = tx.Commit(); err != nil {
-		return nil, fmt.Errorf("committing transaction: %w", err)
-	}
-
-	if err := s.CreateSession(ctx, email, SessionData{
-		Role:        role,
-		Permissions: permissions,
-		Active:      true,
-	}); err != nil {
-		return nil, status.Error(codes.Unavailable, "session store unavailable")
-	}
-
-	return &userpb.RefreshResponse{AccessToken: newAccessToken, RefreshToken: newSignedToken, Permissions: permissions, Role: role}, nil
-}
-
-// getRoleAndPermissions determines the role, permissions, and active status for a user by email.
-// Employees get role "employee" with their DB permissions; clients get role "client" with empty permissions.
-// The active flag is only meaningful for employees; clients always return true.
-func (s *Server) getRoleAndPermissions(email string) (role string, permissions []string, active bool) {
-	emp, err := s.repo.GetEmployeeByAttribute("email", email)
-	if err == nil && emp != nil {
-		permissions := make([]string, len(emp.Permissions))
-		for i, v := range emp.Permissions {
-			permissions[i] = v.Name
-		}
-		return "employee", permissions, emp.Active
-	}
-	return "client", []string{}, true
-}
-
-func (s *Server) Login(ctx context.Context, req *userpb.LoginRequest) (*userpb.LoginResponse, error) {
-	l := logger.FromContext(ctx).With("email", req.Email)
-	user, err := s.repo.GetUserByEmail(req.Email)
-	if err != nil || user == nil {
-		l.WarnContext(ctx, "audit: login failed", "reason", "unknown user")
-		return nil, status.Error(codes.Unauthenticated, "wrong creds")
-	}
-	hashedPassword := HashPassword(req.Password, user.Salt)
-
-	if bytes.Equal(hashedPassword, user.HashedPassword) {
-		role, permissions, active := s.getRoleAndPermissions(user.Email)
-
-		if !active {
-			l.WarnContext(ctx, "audit: login failed", "reason", "deactivated")
-			return nil, status.Error(codes.Unauthenticated, "account deactivated")
-		}
-
-		accessToken, err := s.GenerateAccessToken(user.Email)
-		if err != nil {
-			return nil, err
-		}
-		refreshToken, err := s.GenerateRefreshToken(user.Email)
-		if err != nil {
-			return nil, err
-		}
-		err = s.repo.InsertRefreshToken(refreshToken)
-		if err != nil {
-			return nil, err
-		}
-
-		if err := s.CreateSession(ctx, user.Email, SessionData{
-			Role:        role,
-			Permissions: permissions,
-			Active:      true,
-		}); err != nil {
-			return nil, status.Error(codes.Unavailable, "session store unavailable")
-		}
-
-		l.InfoContext(ctx, "audit: login success", "role", role)
-		return &userpb.LoginResponse{
-			AccessToken:  accessToken,
-			RefreshToken: refreshToken,
-			Permissions:  permissions,
-			Role:         role,
-		}, nil
-	}
-
-	l.WarnContext(ctx, "audit: login failed", "reason", "wrong password")
-	return nil, status.Error(codes.Unauthenticated, "wrong creds")
-}
-
-func (s *Server) Logout(ctx context.Context, req *userpb.LogoutRequest) (*userpb.LogoutResponse, error) {
-	email := req.Email
-	tx, err := s.database.Begin()
-	if err != nil {
-		return nil, fmt.Errorf("starting transaction: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-	err = s.repo.RevokeRefreshTokensByEmail(tx, email)
-	if err != nil {
-		return nil, err
-	}
-
-	if err = tx.Commit(); err != nil {
-		return nil, fmt.Errorf("committing transaction: %w", err)
-	}
-
-	_ = s.DeleteSession(ctx, email)
-
-	logger.FromContext(ctx).InfoContext(ctx, "audit: logout", "email", email)
-	return &userpb.LogoutResponse{
-		Success: true,
-	}, nil
-}
-
-func (s *Server) RequestPasswordReset(ctx context.Context, req *userpb.PasswordActionRequest) (*userpb.PasswordActionResponse, error) {
-	return s.requestPasswordAction(ctx, strings.TrimSpace(req.Email), PasswordActionReset)
-}
-
-func (s *Server) RequestInitialPasswordSet(ctx context.Context, req *userpb.PasswordActionRequest) (*userpb.PasswordActionResponse, error) {
-	return s.requestPasswordAction(ctx, strings.TrimSpace(req.Email), PasswordActionInitialSet)
-}
-
-func (s *Server) SetPasswordWithToken(ctx context.Context, req *userpb.SetPasswordWithTokenRequest) (*userpb.SetPasswordWithTokenResponse, error) {
-	token := strings.TrimSpace(req.Token)
-	newPassword := strings.TrimSpace(req.NewPassword)
-
-	if token == "" || newPassword == "" {
-		return nil, status.Error(codes.InvalidArgument, "token and new password are required")
-	}
-
-	tx, err := s.database.Begin()
-	if err != nil {
-		return nil, status.Error(codes.Internal, "starting transaction failed")
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	email, actionType, err := s.repo.ConsumePasswordActionToken(tx, HashValue(token))
-	if err != nil {
-		if errors.Is(err, repo.ErrInvalidPasswordActionToken) {
-			return nil, status.Error(codes.InvalidArgument, "invalid or expired token")
-		}
-		return nil, status.Error(codes.Internal, "token validation failed")
-	}
-
-	user, err := s.repo.GetUserByEmail(email)
-	if err != nil || user == nil {
-		return nil, status.Error(codes.Internal, "user lookup failed")
-	}
-
-	hashedPassword := HashPassword(newPassword, user.Salt)
-
-	if err := s.repo.UpdatePasswordByEmail(tx, email, hashedPassword); err != nil {
-		return nil, status.Error(codes.Internal, "password update failed")
-	}
-
-	if actionType == PasswordActionInitialSet {
-		if _, err := tx.Exec(`UPDATE employees SET active = true, updated_at = NOW() WHERE email = $1`, email); err != nil {
-			return nil, status.Error(codes.Internal, "employee activation failed")
-		}
-	}
-
-	if err := s.repo.RevokeRefreshTokensByEmail(tx, email); err != nil {
-		return nil, status.Error(codes.Internal, "refresh token revocation failed")
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, status.Error(codes.Internal, "committing transaction failed")
-	}
-
-	_ = s.DeleteSession(ctx, email)
-
-	logger.FromContext(ctx).InfoContext(ctx, "audit: password set via token", "email", email, "action", actionType)
-	return &userpb.SetPasswordWithTokenResponse{Successful: true}, nil
-}
-
-func (s *Server) requestPasswordAction(ctx context.Context, email string, actionType string) (*userpb.PasswordActionResponse, error) {
-	if email == "" {
-		return nil, status.Error(codes.InvalidArgument, "email is required")
-	}
-
-	user, err := s.repo.GetUserByEmail(email)
-
-	if err != nil {
-		return nil, status.Error(codes.Internal, "user lookup failed")
-	}
-	if user == nil {
-		return &userpb.PasswordActionResponse{Accepted: true}, nil
-	}
-
-	token, err := generateOpaqueToken()
-	if err != nil {
-		return nil, status.Error(codes.Internal, "token generation failed")
-	}
-
-	validUntil := time.Now().Add(resetPasswordTokenTTL)
-	if actionType == PasswordActionInitialSet {
-		validUntil = time.Now().Add(initialSetPasswordTTL)
-	}
-
-	if err := s.repo.UpsertPasswordActionToken(user.Email, actionType, HashValue(token), validUntil); err != nil {
-		return nil, status.Error(codes.Internal, "storing token failed")
-	}
-
-	baseURL := os.Getenv("PASSWORD_RESET_BASE_URL")
-	if actionType == PasswordActionInitialSet {
-		baseURL = os.Getenv("PASSWORD_SET_BASE_URL")
-	}
-	link, err := buildActionLink(baseURL, token)
-	if err != nil {
-		return nil, status.Error(codes.Internal, "building password link failed")
-	}
-
-	if err := s.sendPasswordActionEmail(ctx, user.Email, link, actionType); err != nil {
-		return nil, err
-	}
-
-	return &userpb.PasswordActionResponse{Accepted: true}, nil
-}
-
-func (s *Server) sendPasswordActionEmail(ctx context.Context, email string, link string, actionType string) error {
-	notificationAddr := os.Getenv("NOTIFICATION_GRPC_ADDR")
-	if strings.TrimSpace(notificationAddr) == "" {
-		notificationAddr = defaultNotificationURL
-	}
-
-	conn, err := grpc.NewClient(
-		notificationAddr,
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-	)
-	if err != nil {
-		return fmt.Errorf("dialing notification service: %w", err)
-	}
-	defer func() { _ = conn.Close() }()
-
-	client := notificationpb.NewNotificationServiceClient(conn)
-
-	sendCtx, cancelSend := context.WithTimeout(ctx, 5*time.Second)
-	defer cancelSend()
-
-	req := &notificationpb.PasswordLinkMailRequest{
-		ToAddr: email,
-		Link:   link,
-	}
-
-	if actionType == PasswordActionInitialSet {
-		resp, err := client.SendInitialPasswordSetEmail(sendCtx, req)
-		if err != nil {
-			return fmt.Errorf("calling SendInitialPasswordSetEmail: %w", err)
-		}
-		if !resp.Successful {
-			return fmt.Errorf("notification service reported unsuccessful initial set send")
-		}
-		return nil
-	}
-
-	resp, err := client.SendPasswordResetEmail(sendCtx, req)
-	if err != nil {
-		return fmt.Errorf("calling SendPasswordResetEmail: %w", err)
-	}
-	if !resp.Successful {
-		return fmt.Errorf("notification service reported unsuccessful reset send")
-	}
-	return nil
-}
-
-func generateOpaqueToken() (string, error) {
-	random := make([]byte, 32)
-	if _, err := rand.Read(random); err != nil {
-		return "", err
-	}
-	return base64.RawURLEncoding.EncodeToString(random), nil
-}
-
-func HashValue(value string) []byte {
-	sum := sha256.Sum256([]byte(value))
-	return sum[:]
-}
-
-func buildActionLink(baseURL string, token string) (string, error) {
-	if strings.TrimSpace(baseURL) == "" {
-		return "", fmt.Errorf("base URL is empty")
-	}
-
-	parsedURL, err := url.Parse(baseURL)
-	if err != nil {
-		return "", fmt.Errorf("parsing base URL: %w", err)
-	}
-
-	query := parsedURL.Query()
-	query.Set("token", token)
-	parsedURL.RawQuery = query.Encode()
-
-	return parsedURL.String(), nil
-}
-
-func (s *Server) CreateClientAccount(ctx context.Context, req *userpb.CreateClientRequest) (*userpb.CreateClientResponse, error) {
-	is_null := func(str string) bool {
-		return strings.TrimSpace(str) == ""
-	}
-	vals := []string{req.FirstName, req.LastName, req.Gender, req.Email, req.PhoneNumber,
-		req.Address}
-
-	if slices.ContainsFunc(vals, is_null) {
-		return nil, status.Error(codes.InvalidArgument, "One of the required cols is null")
-	}
-
-	if req.Gender != "M" && req.Gender != "F" {
-		return nil, status.Error(codes.InvalidArgument, "Gender must be one of M or F")
-	}
-
-	salt, salt_err := generateSalt()
-	if salt_err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "error generating salt", "err", salt_err)
-		return nil, status.Error(codes.Internal, "Password salting failed")
-	}
-
-	client := model.Client{First_name: req.FirstName,
-		Last_name: req.LastName, Date_of_birth: time.Unix(req.BirthDate, 0),
-		Gender: req.Gender, Email: req.Email, Phone_number: req.PhoneNumber,
-		Address: req.Address, Password: HashPassword(req.Password, salt),
-		Salt_password: salt}
-
-	err := s.repo.CreateClient(client)
-	if err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "client creation failed", "err", err)
-		return nil, status.Error(codes.Internal, "Client creation failed")
-	}
-	return &userpb.CreateClientResponse{Valid: true}, nil
-
-}
-
-func (s *Server) CreateEmployeeAccount(ctx context.Context, req *userpb.CreateEmployeeRequest) (*userpb.GetEmployeeResponse, error) {
-	is_null := func(str string) bool {
-		return strings.TrimSpace(str) == ""
-	}
-	vals := []string{req.FirstName, req.LastName, req.Email,
-		req.Username}
-	if slices.ContainsFunc(vals, is_null) {
-		return nil, status.Error(codes.InvalidArgument, "One of the required cols is null")
-	}
-
-	salt, salt_err := generateSalt()
-	if salt_err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "error generating salt", "err", salt_err)
-	}
-
-	// Spec p.38: every admin is also a supervisor.
-	reqPerms := EnsureAdminImpliesSupervisor(req.Permissions)
-
-	permissions := make([]model.Permission, 0, len(reqPerms))
-	for _, permName := range reqPerms {
-		var perm model.Permission
-		if err := s.db_gorm.First(&perm, "name = ?", permName).Error; err != nil {
-			logger.FromContext(ctx).WarnContext(ctx, "permission not found, skipping", "name", permName)
-			continue
-		}
-		permissions = append(permissions, perm)
-	}
-
-	employee := model.Employee{First_name: req.FirstName,
-		Last_name: req.LastName, Date_of_birth: time.Unix(req.BirthDate, 0),
-		Gender: req.Gender, Email: req.Email, Phone_number: req.PhoneNumber,
-		Address: req.Address, Username: req.Username, Position: req.Position,
-		Department: req.Department, Salt_password: salt,
-		Password: []byte{}, Active: true, Permissions: permissions}
-
-	err := s.repo.CreateEmployee(employee)
-
-	if err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "employee creation failed", "err", err)
-		return nil, status.Error(codes.Internal, "Employee creation failed")
-	}
-
-	// Re-fetch to get the auto-assigned ID and properly loaded permissions
-	created, err := s.repo.GetEmployeeByAttribute("email", employee.Email)
-	if err != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "employee created but failed to fetch", "err", err)
-		return employee.ToProtobuf(), nil
-	}
-
-	// Send activation email so the employee can set their own password
-	_, emailErr := s.RequestInitialPasswordSet(ctx, &userpb.PasswordActionRequest{
-		Email: req.Email,
-	})
-	if emailErr != nil {
-		logger.FromContext(ctx).ErrorContext(ctx, "employee created but activation email failed", "err", emailErr)
-	}
-
-	return created.ToProtobuf(), nil
 }
